requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==3.2.16
djangorestframework==3.12.4
idna==3.10
lxml==5.4.0
PyMuPDF==1.25.5
python-docx==1.1.2
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
